Update draft? to ignore parent_draft? method.
I don’t want to deal with funky file paths for drafts and having to
move files around. That sounded like a good idea at first but its much
easier to just flag from within a post file that its a draft since
someone will be editing it anyway, all the have to do to make it live
is just remove the draft setter call.