Remove post.mtime getter as its unnecessary for public API.
I don’t believe we really need this but if we do anyone can grab it
from the post.path object. For now to keep the API as simple as
possible remove it as we only need the modified file time for the
cache_key which we can still test for successfully.