Remove bundler_args
The Gemfile installs dalli only in production but since we have a
cache.rb in lib (all files in lib are required) that requires dalli
even though we don’t use the class unless in production. I’ll find a
better way of how to solve this. Perhaps it is just best to use dalli
locally but I’d rather not. I have tried to make everything else that
runs in production run locally so I guess it shouldn’t be any different
with dalli.