Update feed test to use builder file and Tilt to render the template.
I’ve removed the static XML file because it contains dates with
timezone info and since we run tests using Travis CI in a different
environment the timezone changes. I didn’t want to mock/stub the Date
object’s xmlschema method since thats not something we own. Best way
for now that I could figure a way around this is to use a builder file
as a fixture. This is probably best and easier to edit since it’ll
dynamically create data from the test data we have.